requirements.txt
console-menu==0.8.0
ansiwrap==0.8.4
textwrap3==0.9.2
Pillow==10.0.0
fpdf2==2.7.6
python-docx==0.8.11